backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-multipart>=0.0.12
httpx>=0.27.0
pydantic>=2.9.0
pydantic-settings>=2.6.0
pymupdf>=1.24.0
chromadb>=0.5.0
sentence-transformers>=3.0.0
rank-bm25>=0.2.2
numpy>=1.26.0
pytesseract>=0.3.10
Pillow>=10.0.0
